ddb: allow custom types for hash and range keys as long as they serialize as numbers or strings;
Previously, we tested for each individual type we supported directly and
thus only allowed a closed set of types for ddb hash and range fields.
This only included the primitive types and time.Time, so if you needed
anything else you were out of luck. Instead, we are now looking at how a
type serializes as JSON to make this decision.