db-repo, apiserver: allow string primary keys and added some generics;
This commit changes the way you work with `Repository`s and CRUD
handlers. For repositories, we can now use strings as primary keys as
well (using `DistributedModel`s) and the methods to access entities are
now tied to the actual entities (giving you more type safety). The
repository now returns you the correct values directly, no need to pass
an empty value into the repository anymore.

The CRUD handlers are now also fully typed (so if you have `FooCreateInput`
as your create input, that is exactly the type you are getting in the
`TransformCreate` callback, no manual type casting needed). The
`GetCreateInput` and similar methods have also been removed as we can
now just create a zero value as needed.

** This is a breaking change. **